# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "Used to manage interactions between customers, mechanics, and service tickets in a typical Mechanic Shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: 
...
  /mechanics/login:
    post:
      tags:
        - mechanics
      summary: "Login endpoint, (generates token)"
      description: "Validates mechanic credentials, generates and return a mechanic specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
  /mechanics: 
    post: 
      tags:
        - mechanics
      summary: "Endpoint to create/register a new mechanic."
      description: "Endpoint to create a new mechanic in our mechanic shop system, emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new mechanic."
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"
      responses:
        201:
          descriptions: "Successful Mechanic Creation"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              email: johnd@email.com
              phone: "123-456-7890"
              salary: 30.000
              password: password123
          
    get: #get all mechanics endpoint
      tags:
        - mechanics
      summary: "Returns all Mechanics"
      description: "Endpoint to retrieve a list of all mechanics."
      responses:
        200:
          description: "Retrieved Mechanics Successfully"           
          schema:
            $ref: "#/definitions/AllMechanics" 
          
    put: 
      tags:
        - mechanics
      summary: "Endpoint to update a mechanic."
      description: "Endpoint to update a mechanic, this is a token authenticated route, and the mechanic must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a Mechanic."
          required: true
          schema:
            $ref: "#/definitions/UpdateMechanicInput"
      responses:
        200:
          descriptions: "Successfully Updated mechanic"
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"
          examples:
            application/json:
              id: 1
              name: Dohn Joe
              phone: "123-456-7890"
              email: dohnj@email.com
              salary: 30000
              password: password123 

    delete:
      tags:
        - mechanics
      summary: "Endpoint to delete a mechanic"
      description: "Endpoint to delete a mechanic. Requires token authentication, mechanic must be logged in to delete their account"
      security:
       - bearerAuth: []
      responses:
        200:
          description: "Successfully Deleted mechanic"           
          schema:
            $ref: "#/definitions/DeleteMechanicResponse"
          examples:
            application/json:
              message: "successfully deleted mechanic <mechanic_id>."
  /mechanics/my-tickets:
    get: #get all mechanics tickets endpoint
      tags:
        - mechanics
      summary: "Returns all tickets for a mechanic"
      description: "Endpoint to retrieve a list of all tickets for a mechanic."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Retrieved Tickets Successfully"           
          schema:
            $ref: "#/definitions/TicketResponse" 

# SERVICE TICKETS
...
definitions: 
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
      
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
...
  MechanicInput:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "float"
      password:
        type: "string"
      
    required:
      - name
      - email
      - phone
      - salary
      - password
  MechanicResponse:
    type: "object"
    properties:
      id:
        type: "int"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary: 
        type: "float"
      password:
        type: "string"
  AllMechanics:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        salary:
          type: "float"
        password:
          type: "string"
...
  TicketResponse:
    type: "object"
    properties:
      id:
        type: "int"
      VIN:
        type: "string"
      service_date:
        type: "string"
      service_desc:
        type: "string"
      customer_id: 
        type: "int"
...